polar: Add OOT for polar codes
This OOT is a wrapper for `polar-codes` a library with optimized polar
encoders and decoders. It depends on some more libraries that are not
yet available with the standard PyBOMBS recipes.
fmt.lwr installes FMTlib which may be very useful to move towards
`std::format` and purge `boost::format`.
tclap is a tool to create a command line interface in C++.